requirements.txt
google-generativeai
pinecone>=3.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0
streamlit>=1.28.0
python-dotenv>=1.0.0
lxml[html_clean]>=4.9.0
selenium>=4.15.0
webdriver-manager>=4.0.0
